Introduce a shallow ground ordering generic function
This comes along with a warning/clarification to the assoccomm relations.
They have always walked the first argument, which, when used with the
old `ground_order` would only account for improperly ordered terms when
the second argument was fully ground (and the first wasn't).  However, it
would perform this check repeatedly, incurring a full-traversal cost
on every iteration.  This inefficiency has been rectified.

The new "shallow" ground order should accomplish the same--and more--by
reordering according to "groundness" only at next level of traversal.
The assoccomm goals need to be updated so that they properly use this
shallow ordering, though.